+++ src/cli/subcommands/send_cmd.rs
@@ -3,6 +3,8 @@
 
 use std::str::FromStr as _;
 
+use crate::rpc::mpool_api::MpoolPushMessage;
+use crate::rpc::{self, RpcMethodExt as _};
 use crate::rpc_client::ApiInfo;
 use crate::shim::address::{Address, StrictAddress};
 use crate::shim::econ::TokenAmount;
@@ -32,6 +34,11 @@ pub struct SendCommand {
 
 impl SendCommand {
     pub async fn run(self, api: ApiInfo) -> anyhow::Result<()> {
+        eprintln!(
+            "This command has been deprecated and will be removed in the future.\n\
+             Please use the 'forest-wallet' executable instead."
+        );
+
         let from: Address =
             if let Some(from) = &self.from {
                 StrictAddress::from_str(from)?.into()
@@ -53,7 +60,9 @@ impl SendCommand {
             ..Default::default()
         };
 
-        let signed_msg = api.mpool_push_message(message, None).await?;
+        let signed_msg = MpoolPushMessage::call(&rpc::Client::from(api), (message.into(), None))
+            .await?
+            .into_inner();
 
         println!("{}", signed_msg.cid().unwrap());