+++ src/lotus_json/mod.rs
@@ -126,6 +126,7 @@ use derive_more::From;
 use fil_actor_interface::{miner::DeadlineInfo, power::Claim};
 use fil_actors_shared::fvm_ipld_bitfield::json::BitFieldJson;
 use fil_actors_shared::fvm_ipld_bitfield::BitField;
+use fvm_shared2::piece::PaddedPieceSize;
 use schemars::{gen::SchemaGenerator, schema::Schema, JsonSchema};
 use serde::{de::DeserializeOwned, Deserialize, Deserializer, Serialize, Serializer};
 #[cfg(test)]
@@ -413,7 +414,9 @@ where
 }
 
 /// A domain struct that is (de) serialized through its lotus JSON representation.
-#[derive(Debug, Serialize, Deserialize, From, Default, Clone)]
+#[derive(
+    Debug, Serialize, Deserialize, From, Default, Clone, PartialEq, Eq, PartialOrd, Ord, Hash,
+)]
 #[serde(bound = "T: HasLotusJson + Clone")]
 pub struct LotusJson<T>(#[serde(with = "self")] pub T);
 
@@ -497,6 +500,7 @@ lotus_json_with_self!(
     std::path::PathBuf,
     bool,
     DeadlineInfo,
+    PaddedPieceSize,
 );
 
 #[derive(Default, Debug, Serialize, Deserialize)]