+++ src/rpc/state_api.rs
@@ -13,7 +13,7 @@ use crate::shim::{
     state_tree::ActorState, version::NetworkVersion,
 };
 use crate::state_manager::chain_rand::ChainRand;
-use crate::state_manager::vm_circ_supply::GenesisInfo;
+use crate::state_manager::circulating_supply::GenesisInfo;
 use crate::state_manager::{InvocResult, MarketBalance};
 use crate::utils::db::car_stream::{CarBlock, CarWriter};
 use ahash::{HashMap, HashMapExt};
@@ -25,7 +25,7 @@ use fil_actor_interface::miner::DeadlineInfo;
 use fil_actor_interface::{
     market, miner,
     miner::{MinerInfo, MinerPower},
-    multisig, power,
+    multisig, power, reward,
 };
 use fil_actors_shared::fvm_ipld_bitfield::BitField;
 use futures::StreamExt;
@@ -35,7 +35,9 @@ use jsonrpsee::types::{error::ErrorObject, Params};
 use libipld_core::ipld::Ipld;
 use nonempty::{nonempty, NonEmpty};
 use num_bigint::BigInt;
+use num_traits::Euclid;
 use parking_lot::Mutex;
+use std::ops::Mul;
 use std::path::PathBuf;
 use std::{sync::Arc, time::Duration};
 use tokio::task::JoinSet;
@@ -77,6 +79,8 @@ pub const STATE_MINER_SECTOR_COUNT: &str = "Filecoin.StateMinerSectorCount";
 pub const STATE_VERIFIED_CLIENT_STATUS: &str = "Filecoin.StateVerifiedClientStatus";
 pub const STATE_VM_CIRCULATING_SUPPLY_INTERNAL: &str = "Filecoin.StateVMCirculatingSupplyInternal";
 pub const STATE_MARKET_STORAGE_DEAL: &str = "Filecoin.StateMarketStorageDeal";
+pub const STATE_DEAL_PROVIDER_COLLATERAL_BOUNDS: &str =
+    "Filecoin.StateDealProviderCollateralBounds";
 pub const MSIG_GET_AVAILABLE_BALANCE: &str = "Filecoin.MsigGetAvailableBalance";
 pub const MSIG_GET_PENDING: &str = "Filecoin.MsigGetPending";
 
@@ -860,8 +864,11 @@ pub async fn state_circulating_supply<DB: Blockstore + Send + Sync + 'static>(
 
     let genesis_info = GenesisInfo::from_chain_config(state_manager.chain_config());
 
-    let supply =
-        genesis_info.get_circulating_supply(height, &state_manager.blockstore_owned(), root)?;
+    let supply = genesis_info.get_state_circulating_supply(
+        height,
+        &state_manager.blockstore_owned(),
+        root,
+    )?;
 
     Ok(LotusJson(supply))
 }
@@ -1072,10 +1079,72 @@ pub async fn state_market_storage_deal<DB: Blockstore + Send + Sync + 'static>(
         .context("Market actor not found")?;
     let market_state = market::State::load(store, actor.code, actor.state)?;
     let proposals = market_state.proposals(store)?;
-    let proposal =  proposals.get(deal_id)?.ok_or_else(|| anyhow::anyhow!("deal {deal_id} not found - deal may not have completed sealing before deal proposal start epoch, or deal may have been slashed"))?;
+    let proposal = proposals.get(deal_id)?.ok_or_else(|| anyhow::anyhow!("deal {deal_id} not found - deal may not have completed sealing before deal proposal start epoch, or deal may have been slashed"))?;
 
     let states = market_state.states(store)?;
     let state = states.get(deal_id)?.unwrap_or_else(DealState::empty);
 
     Ok(MarketDeal { proposal, state }.into())
+}
+pub async fn state_deal_provider_collateral_bounds<DB: Blockstore + Send + Sync + 'static>(
+    params: Params<'_>,
+    data: Ctx<DB>,
+) -> Result<DealCollateralBounds, JsonRpcError> {
+    let deal_provider_collateral_num = BigInt::from(110);
+    let deal_provider_collateral_denom = BigInt::from(100);
+
+    let LotusJson((size, verified, ApiTipsetKey(tsk))) = params.parse()?;
+
+    // This is more eloquent than giving the whole match pattern a type.
+    let _: bool = verified;
+
+    let state_manager = &data.state_manager;
+    let ts = state_manager
+        .chain_store()
+        .load_required_tipset_or_heaviest(&tsk)?;
+
+    let power_actor = state_manager
+        .get_actor(&Address::POWER_ACTOR, *ts.parent_state())?
+        .context("Power actor address could not be resolved")?;
+
+    let reward_actor = state_manager
+        .get_actor(&Address::REWARD_ACTOR, *ts.parent_state())?
+        .context("Power actor address could not be resolved")?;
+
+    let store = state_manager.blockstore();
+
+    let power_state = power::State::load(store, power_actor.code, power_actor.state)?;
+    let reward_state = reward::State::load(store, reward_actor.code, reward_actor.state)?;
+
+    let genesis_info = GenesisInfo::from_chain_config(state_manager.chain_config());
+
+    let supply = genesis_info.get_vm_circulating_supply(
+        ts.epoch(),
+        &data.state_manager.blockstore_owned(),
+        ts.parent_state(),
+    )?;
+
+    let power_claim = power_state.total_power();
+
+    let policy = &state_manager.chain_config().policy;
+
+    let baseline_power = reward_state.this_epoch_baseline_power();
+
+    let (min, max) = reward_state.deal_provider_collateral_bounds(
+        policy,
+        size,
+        &power_claim.raw_byte_power,
+        baseline_power,
+        &supply.into(),
+    );
+
+    let min = min
+        .atto()
+        .mul(deal_provider_collateral_num)
+        .div_euclid(&deal_provider_collateral_denom);
+
+    Ok(DealCollateralBounds {
+        max: max.into(),
+        min: TokenAmount::from_atto(min),
+    })
 }