+++ CHANGELOG.md
@@ -29,11 +29,17 @@
 
 ### Added
 
+- [#4084](https://github.com/ChainSafe/forest/pull/4084) Add support for the
+  `Filecoin.StateDealProviderCollateralBounds` RPC method.
+
 ### Changed
 
 ### Removed
 
 ### Fixed
+
+- [#4145](https://github.com/ChainSafe/forest/pull/4145) Fix the
+  `forest-cli net peers --agent` command in case the agent is not available.
 
 ## Forest 0.17.1 "Villentretenmerth"
 
@@ -109,6 +115,12 @@ Mandatory release that includes:
 
 - [#3955](https://github.com/ChainSafe/forest/pull/3955) Added support for the
   NV22 _Dragon_ network upgrade, together with the required state migration.
+
+### Changed
+
+- [#3976](https://github.com/ChainSafe/forest/pull/3976) `forest-wallet`
+  defaults to using a local wallet instead of the builtin Forest wallet for
+  greater security.
 
 ### Fixed